requirements.txt
asgiref==3.3.4
Django==3.2.3
django-redis==5.0.0
djangorestframework==3.12.4
mysqlclient==2.0.3
PyMySQL==1.0.2
pytz==2021.1
redis==3.5.3
sqlparse==0.4.1
